add date_square and date_accurate
Added "days_accurate" format variable passed to provide_version() to accurately indicate the number of days since the milestone.

This change uses DateTime::delta_days(). Note that delta_md() is also
available as a method to calculate the number of months and days since the
milestone, so it is possible as an enhancement to also provide the new format
variables "months" and "mdays".

Will replace days